ref(admin): Refactor Runtime notification client
